Cleanup alp example scripts
